Add IFittingModel interface and related structs/enums
Added a new interface `IFittingModel` in the `TAFitting.Model`
namespace. This interface represents a fitting model and
includes properties for the model's name, description, and
parameters, as well as a method to get a function based on
the current model with specified parameters.

Introduced a new `Parameter` struct in the `TAFitting.Model`
namespace. This struct represents a fitting parameter and
includes properties for the parameter's name, constraints,
and initial value.

Added a new `ParameterConstraints` enum in the `TAFitting.Model`
namespace. This enum represents the constraints of a parameter
and includes flags for no constraints, integer constraints,
positive constraints, and non-negative constraints.