Add ExcelFormula property to IFittingModel interface
A new property `ExcelFormula` was added to the `IFittingModel`
interface. This property is a string that provides the formula
of the model for use in Excel. The formula has specific notation
requirements: parameter names must be enclosed in square brackets
(e.g., "[A0]"), time is denoted as "$X", and an equal sign is not
required. An example formula for a single-component exponential
model is provided: "[A0] + [A1] * EXP(-$X / [T1])".