Add SpectraMarkerSize property to AppearanceConfig
A new property `SpectraMarkerSize` has been added to the `AppearanceConfig` class in `AppearanceConfig.cs`. This property is used to get or set the size of the spectra marker and is initialized to 7.

In `SpectraPreviewWindow.cs`, the `markerSize` field has been updated to use the new `Program.SpectraMarkerSize` property instead of a hardcoded value of 7. The `ChangeMarkerSize` method has been modified to update `Program.SpectraMarkerSize` when the marker size is changed.

A new static property `SpectraMarkerSize` has been added to the `Program` class in `Program.cs`. This property gets or sets the `SpectraMarkerSize` from the `AppearanceConfig` and ensures the configuration is saved whenever the value is changed.