Load decay data asynchronously to improve UI responsiveness
The code change modifies the loading of decay data from a folder to run asynchronously. Previously, the `Decays.FromFolder` method and subsequent UI updates were executed on the main thread. The updated code now uses `Task.Run` to perform the data loading in a background task. Once the data is loaded, the UI updates (setting `rangeSelector.Time.To` and calling `MakeTable`) are invoked on the main thread using `Invoke`.

These changes improve the responsiveness of the application by offloading potentially time-consuming operations to a background thread, preventing the UI from freezing during the data loading process.