Add IsMagnitude property to Parameter struct
The `Parameter` struct in `Parameter.cs` now includes a new boolean property `IsMagnitude`, which indicates if the parameter represents a magnitude. If `IsMagnitude` is set to `true`, the parameter will be inverted automatically in inverted mode.

In `EmpiricalPowerLaw.cs`, the `IsMagnitude` property is set to `true` for the parameter named "A0".

In `Exponential1Component.cs`, the `IsMagnitude` property is set to `true` for the parameters named "A0" and "A1".