Enhance ModelManager to load models from "models" directory
The static constructor of the `ModelManager` class has been
modified to load additional model assemblies from a directory
named "models" located in the application's root directory.
This is achieved by iterating over all `.dll` files in the
"models" directory and attempting to load each one. If an
exception occurs during the loading of a file, it is caught
and written to the debug output.

These changes make the system more flexible and extensible,
allowing new models to be added simply by placing their
assemblies in the "models" directory without modifying the code.