Add RowsRemoved event handler to update spectra visualization
The code modification adds an event handler for the `RowsRemoved`
event of the `timeTable` object. This event handler is set to the
`DrawSpectra` method, ensuring that the `DrawSpectra` method is
invoked not only when cell values change but also when rows are
removed from the `timeTable`. This helps keep the spectra
visualization updated in response to row deletions.