Lowered minimum values for time and ΔuOD range selectors
The `ToMinimum` value for the `time` range selector has been
changed from `5M` to `0.01M`, allowing the minimum value to be
set as low as `0.01` microseconds instead of `5` microseconds.

The `ToMinimum` value for the `ΔuOD` range selector has been
changed from `50M` to `1M`, allowing the minimum value to be
set as low as `1` instead of `50`.