Add CustomNumericUpDown for Time0 with labels and event
Added a new CustomNumericUpDown control named `nud_time0` to the MainWindow class for setting the `Time0` property of the `decays` object. This control allows input of decimal values with three decimal places, ranging from -1000 to 1000. Added labels "Time 0" and "µs" next to the control for clarity. Reduced the size of the `cb_invert` checkbox to make space for the new controls. Hooked up the `ValueChanged` event of `nud_time0` to a new method `ChangeTime0`, which updates the `Time0` property and refreshes the plots. Set `nud_time0` value to `Time0` when loading decays from a folder.