Add unsaved changes check in LoadDecays method
The `LoadDecays` method in `MainWindow.cs` now includes a user prompt to warn about potential data loss if there are unsaved changes in the `parametersTable`. This is achieved by checking the new `Edited` property of `parametersTable` and displaying a `MessageBox` if any rows have been edited. The method only proceeds to load new data if the user confirms.

The `ParametersTable` class in `ParametersTable.cs` has been enhanced with a new `Edited` property. This property checks if any rows in the `ParameterRows` collection have been marked as edited, providing a way to determine if there are unsaved changes.