Add remarks to clarify GetVectorizedDerivatives behavior
Added a <remarks> section to the XML documentation of the
GetVectorizedDerivatives method in the IVectorizedModel<TVector>
interface. The remarks explain that the second argument of the
returned Action<TVector, TVector[]>, which will contain the
derivatives at the end of the method, is not initialized to 0
when passed to the method. It suggests calling TVector.Load(double)
to fill the array with 0 if necessary.