Add XML documentation to StatsUtils and TDist classes
Introduced new XML documentation comments to the `StatsUtils` and
`TDist` classes in the `TAFitting.Stats` namespace. These comments
provide summaries and descriptions for the methods and classes,
improving code readability and maintainability.

In `StatsUtils.cs`, added documentation for:
- `Variance`: Calculates the variance of a sequence of double-precision
  floating-point numbers.
- `StandardDeviation`: Calculates the standard deviation of a sequence
  of double-precision floating-point numbers.
- `SmirnovGrubbs`: Removes outliers from a sequence of double-precision
  floating-point numbers using the Smirnov-Grubbs test.

In `TDist.cs`, added documentation for the `TDist` class, describing it
as representing a t-distribution.