Add check for non-negative ParallelThreshold
The condition in the `if` statement has been modified to include an additional check. Previously, the code only checked if `source.Length` was greater than or equal to `Program.ParallelThreshold`. Now, it also checks if `Program.ParallelThreshold` is greater than or equal to 0. This enables to configure the app to never parallelize the solving logic.