Add filter application and default setting functionality
This update introduces new menu items to apply filters to selected rows and all rows in the data grid. A new option to set a filter as the default is added, along with a separator for improved menu organization. The `SetDefaultFilter` method is implemented to update the `Program.DefaultFilter` property. Additionally, the `DefaultFilter` property in the `FilterManager` class is modified to use the new default filter setting.