Refactor DataPoint creation for performance
Refactored the `AddDecay` method in `UIUtils.cs` to use a new helper method, `GetOrCreateDataPoint`, for creating or reusing `DataPoint` objects. This method improves performance by avoiding unnecessary calls to the `Invalidate` method when setting the `IsEmpty` property. Added the `[MethodImpl(MethodImplOptions.AggressiveInlining)]` attribute to the helper method for further optimization. Updated comments to explain the rationale behind these changes. Removed explicit setting of `IsEmpty` to `false` in favor of the new method.